Gather cli tests into tests/test_cli.py
